websocket-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
docker==7.1.0
google-cloud-firestore==2.20.0
python-multipart==0.0.19